feat(nft-card): update sale form
