+++ apps/ui/src/modules/profile/components/nft/NftCardSaleForm.tsx
@@ -1,10 +1,12 @@
-import { CurrencyInput } from '@/components';
+import { Card, CurrencyInput } from '@/components';
 import { CloseIcon } from '@/components/icons/closeIcon';
 import type { Asset } from '@/types/marketplace';
 import { ETH_ID } from '@/utils/constants';
 import { ChevronDownIcon } from '@chakra-ui/icons';
 import {
   Button,
+  Divider,
+  Flex,
   FormControl,
   FormErrorMessage,
   FormLabel,
@@ -15,9 +17,14 @@ import {
   Input,
   InputGroup,
   InputRightElement,
+  Stack,
+  Text,
   useDisclosure,
 } from '@chakra-ui/react';
+import { Link } from '@tanstack/react-router';
+import { bn } from 'fuels';
 import { isNaN as lodashIsNaN } from 'lodash';
+import { useMemo } from 'react';
 import { Controller, useForm } from 'react-hook-form';
 import { NftSearchAssetModal } from './NftSearchAssetModal';
 
@@ -33,27 +40,38 @@ export type NftSaleCardForm = {
 interface NftCardSaleFormProps {
   initialValues?: NftSaleCardForm;
   onSubmit: (data: NftSaleCardForm) => void;
-  isLoading?: boolean;
   assets: Asset[];
+  userWithHandle: boolean;
 }
 
 export const NftCardSaleForm = ({
   initialValues,
   onSubmit,
-  isLoading = false,
   assets,
+  userWithHandle,
 }: NftCardSaleFormProps) => {
   const {
     control,
     formState: { errors },
     handleSubmit,
     setValue,
+    watch,
   } = useForm<NftSaleCardForm>({
     defaultValues: initialValues,
   });
   const { isOpen, onOpen, onClose } = useDisclosure();
 
-  const findDefaultAsset = assets.find((asset) => asset.id === ETH_ID);
+  const sellAssetId = watch('sellAsset.id');
+
+  const currentSellAsset = useMemo(
+    () => assets.find((asset) => asset.id === sellAssetId),
+    [assets, sellAssetId]
+  );
+
+  const findDefaultAsset = useMemo(
+    () => assets.find((asset) => asset.id === ETH_ID),
+    [assets]
+  );
   const defaultAsset = {
     id: findDefaultAsset?.id,
     name: findDefaultAsset?.metadata?.name,
@@ -64,9 +82,47 @@ export const NftCardSaleForm = ({
     setValue('sellPrice', 0);
   };
 
+  const userWithHandleFee = useMemo(() => {
+    if (!currentSellAsset) return '0';
+    return currentSellAsset.fees[1];
+  }, [currentSellAsset]);
+
+  const userWithoutHandleFee = useMemo(() => {
+    if (!currentSellAsset) return '0';
+    return currentSellAsset.fees[0];
+  }, [currentSellAsset]);
+
+  const currentFee = useMemo(() => {
+    if (!currentSellAsset) return 0;
+    return userWithHandle
+      ? userWithHandleFee // User with handle pays the second fee
+      : userWithoutHandleFee; // User without handle pays the first fee
+  }, [
+    currentSellAsset,
+    userWithHandle,
+    userWithHandleFee,
+    userWithoutHandleFee,
+  ]);
+
+  const currentValue = watch('sellPrice');
+
+  const valueToReceive = useMemo(() => {
+    if (!currentSellAsset || !currentValue) return 0;
+
+    const valueInBaseUnits = bn.parseUnits(
+      currentValue.toString(),
+      currentSellAsset.metadata?.decimals || 9
+    );
+    const feeAmount = valueInBaseUnits.mul(currentFee).div(10000);
+    const valueAfterFee = valueInBaseUnits.sub(feeAmount);
+
+    return valueAfterFee.formatUnits(currentSellAsset.metadata?.decimals || 9);
+  }, [currentValue, currentSellAsset, currentFee]);
+
   return (
     <Grid
       as="form"
+      id="nft-sale-form"
       onSubmit={handleSubmit(onSubmit)}
       templateColumns="repeat(2, 1fr)"
       gap={4}
@@ -205,15 +261,67 @@ export const NftCardSaleForm = ({
       </GridItem>
 
       <GridItem colSpan={2}>
-        <Button
-          variant="primary"
-          type="submit"
-          disabled={isLoading}
-          isLoading={isLoading}
-        >
-          List NFT
-        </Button>
+        <Card p={4}>
+          <Stack spacing={4}>
+            <Flex justifyContent="space-between" alignItems="center">
+              <Text fontSize="xs" color="grey.subtitle">
+                Price
+              </Text>
+              <Text fontSize="xs" color="grey.100">
+                {currentValue} {currentSellAsset?.metadata?.symbol}
+              </Text>
+            </Flex>
+
+            <Flex justifyContent="space-between" alignItems="center">
+              <Text fontSize="xs" color="grey.subtitle">
+                Bako fee
+              </Text>
+              <Text fontSize="xs" color="grey.100">
+                {`${(Number(currentFee) / 100).toFixed(2)}%`}
+              </Text>
+            </Flex>
+
+            <Divider />
+
+            <Flex justifyContent="space-between" alignItems="center">
+              <Text fontSize="xs" color="grey.subtitle">
+                You will receive
+              </Text>
+              <Text fontSize="xs" color="grey.100">
+                {Number(valueToReceive) || '0'}{' '}
+                {currentSellAsset?.metadata?.symbol}
+              </Text>
+            </Flex>
+          </Stack>
+        </Card>
       </GridItem>
+
+      {!userWithHandle && (
+        <GridItem colSpan={2}>
+          <Card
+            p={4}
+            flexDir="row"
+            alignItems="center"
+            justifyContent="space-between"
+          >
+            <Stack spacing={2}>
+              <Text
+                fontSize="md"
+                fontWeight="bold"
+              >{`Unlock ${bn(userWithHandleFee).formatUnits(2)}% fee`}</Text>
+              <Text fontSize="sm" color="section.200">
+                Turn your address into a Handle <br /> and get{' '}
+                <Text as="span" color="button.500">
+                  {`${bn(userWithHandleFee).formatUnits(2)}%`} trade fees
+                </Text>
+              </Text>
+            </Stack>
+            <Button variant="secondary" w="fit-content" as={Link} to="/">
+              Get now
+            </Button>
+          </Card>
+        </GridItem>
+      )}
     </Grid>
   );
 };
